backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg[binary]==3.1.13
redis==5.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
alembic==1.13.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
# Barcode scanning and image processing
pillow>=10.0.0
pyzbar>=0.1.9
opencv-python-headless>=4.8.0
# WebSocket support
websockets>=12.0
# Notification services
twilio>=8.10.0
sendgrid>=6.10.0
celery>=5.3.0
redis>=5.0.1
schedule>=1.2.0